javascript読み込みエラーpayjp以外対処
