+++ app/javascript/packs/application.js
@@ -14,6 +14,7 @@ require("top_page.js")
 require("detail.js")
 require("profile_edit")
 require("pay")
+require("message.js")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
